migrate `Side`
