+++ packages/styles/theme/_variables.scss
@@ -4,3 +4,4 @@ SPDX-License-Identifier: GPL-3.0-only */
 $page-width-medium-threshold: 1150px;
 $page-width-small-threshold: 826px;
 $row-section-thirds-threshold: 1400px;
+$side-menu-minimised-width: 75px;
