+++ packages/ui-structure/src/Side/index.module.scss
@@ -1,0 +1,36 @@
+// Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+@import 'styles/theme/_variables';
+
+.side {
+  /* maximised by default, or minimised otherwise. */
+  min-width: var(--core-side-width);
+  max-width: var(--core-side-width);
+  z-index: 7;
+  position: sticky;
+  top: 0;
+  height: 100vh;
+  flex: 0;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+  transition: all 0.5s cubic-bezier(0.1, 1, 0.2, 1);
+
+  @media (max-width: $page-width-medium-threshold) {
+    position: fixed;
+    top: 0;
+    left: 0;
+  }
+}
+
+.sideMinimised {
+  min-width: $side-menu-minimised-width;
+  max-width: $side-menu-minimised-width;
+}
+
+.sideHidden {
+  @media (max-width: $page-width-medium-threshold) {
+    left: calc(var(--core-side-width) * -1);
+  }
+}
