+++ packages/ui-structure/src/index.tsx
@@ -11,3 +11,4 @@ export * from './PageHeading';
 export * from './PageTitle';
 export * from './PageTitleTabs';
 export * from './RowSection';
+export * from './Side';
